mppf-casos-fiscales/ETL/requirements.txt
pandas
python-dateutil
numpy
